cleanup

+++ lib/macaw/rules/clean.rb
@@ -3,8 +3,9 @@
 # requires arara 3.0+
 class Macaw
   def clean(parameters)
-    sleep(parameters.wait || 0)
-    (parameters.files || []).reject{|f| f.downcase == @file.downcase}.each{|f|
+    return unless parameters.files
+    sleep(parameters.wait) if paramaters.wait.to_i != 0
+    parameters.files.reject{|f| f.downcase == @file.downcase}.each{|f|
       File.unlink(f)
     }
   end
